requirements/base.txt
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-camel-case==1.3.0
drf-yasg==1.20.0
django-cors-headers==3.13.0
requests==2.28.1
boto3==1.24.45

python-dotenv==0.21.0
dj-database-url==1.0.0

gunicorn==20.1.0
whitenoise==6.2.0
redis==4.3.2
celery==5.2.7
gevent==21.12.0
